# apps/delivery-service/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/auth/register": {
      "post": {
        "operationId": "AuthController_register",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "User registered successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Validation error"
          },
          "409": {
            "description": "Email already exists"
          }
        },
        "summary": "Register a new user [Public]",
        "tags": [
          "Auth"
        ]
      }
    },
    "/api/auth/login": {
      "post": {
        "operationId": "AuthController_login",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Login successful",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Invalid credentials"
          }
        },
        "summary": "Login user [Public]",
        "tags": [
          "Auth"
        ]
      }
    },
    "/api/auth/refresh": {
      "post": {
        "operationId": "AuthController_refreshToken",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RefreshTokenDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Token refreshed successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Invalid refresh token"
          }
        },
        "summary": "Refresh access token [Public]",
        "tags": [
          "Auth"
        ]
      }
    },
    "/api/auth/me": {
      "get": {
        "operationId": "AuthController_me",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Current user info",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserInfoResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get current user info [Any Role]",
        "tags": [
          "Auth"
        ]
      }
    },
    "/api/auth/users/{userId}/role": {
      "patch": {
        "operationId": "AuthController_updateUserRole",
        "parameters": [
          {
            "name": "userId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateUserRoleDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Role updated successfully"
          },
          "403": {
            "description": "Forbidden - Admin only"
          },
          "404": {
            "description": "User not found"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Update user role [Admin]",
        "tags": [
          "Auth"
        ]
      }
    },
...
    "/api/cart": {
      "get": {
        "operationId": "CartsController_getMyCart",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Cart details",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CartResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get current customer cart [Customer]",
        "tags": [
          "Carts"
        ]
      },
      "delete": {
        "operationId": "CartsController_clearCart",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Cart cleared"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Clear cart [Customer]",
        "tags": [
          "Carts"
        ]
      }
    },
    "/api/cart/items": {
      "post": {
        "operationId": "CartsController_addItem",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AddItemToCartDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Item added to cart"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Add item to cart [Customer]",
        "tags": [
          "Carts"
        ]
      },
      "patch": {
        "operationId": "CartsController_updateItemQuantity",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateItemQuantityDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Item quantity updated"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Update item quantity in cart [Customer]",
        "tags": [
          "Carts"
        ]
      },
      "delete": {
        "operationId": "CartsController_removeItem",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RemoveItemFromCartDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Item removed from cart"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Remove item from cart [Customer]",
        "tags": [
          "Carts"
        ]
      }
    },
    "/api/cart/checkout": {
      "post": {
        "operationId": "CartsController_checkout",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CheckoutCartDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Order created from cart"
          },
          "400": {
            "description": "Cart is empty"
          },
          "404": {
            "description": "Cart not found"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Checkout cart and create order [Customer]",
        "tags": [
          "Carts"
        ]
      }
    },
...
  },
  "info": {
    "title": "Delivery Service API",
    "description": "A modular monolith delivery service system with Clean Architecture, CQRS, and Event-Driven patterns",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
      "RegisterDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "format": "email"
          },
          "password": {
            "type": "string",
            "minLength": 8
          },
          "name": {
            "type": "string",
            "minLength": 2
          },
          "phone": {
            "type": "string",
            "minLength": 9
          }
        },
        "required": [
          "email",
          "password",
          "name",
          "phone"
        ]
      },
      "AuthResponseDto": {
        "type": "object",
        "properties": {
          "accessToken": {
            "type": "string",
            "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
          },
          "refreshToken": {
            "type": "string",
            "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
          }
        },
        "required": [
          "accessToken",
          "refreshToken"
        ]
      },
      "LoginDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "format": "email"
          },
          "password": {
            "type": "string",
            "minLength": 1
          }
        },
        "required": [
          "email",
          "password"
        ]
      },
      "RefreshTokenDto": {
        "type": "object",
        "properties": {
          "refreshToken": {
            "type": "string",
            "minLength": 1
          }
        },
        "required": [
          "refreshToken"
        ]
      },
      "UserInfoResponseDto": {
        "type": "object",
        "properties": {
          "userId": {
            "type": "string",
            "example": "123e4567-e89b-12d3-a456-426614174000"
          },
          "email": {
            "type": "string",
            "example": "user@example.com"
          },
          "roles": {
            "example": [
              "customer"
            ],
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "required": [
          "userId",
          "email",
          "roles"
        ]
      },
      "UpdateUserRoleDto": {
        "type": "object",
        "properties": {
          "role": {
            "type": "string",
            "enum": [
              "customer",
              "driver",
              "admin",
              "restaurant_owner"
            ]
          }
        },
        "required": [
          "role"
        ]
      },
...
      "CartItemDto": {
        "type": "object",
        "properties": {
          "menuItemId": {
            "type": "string",
            "example": "123e4567-e89b-12d3-a456-426614174000"
          },
          "restaurantId": {
            "type": "string",
            "example": "123e4567-e89b-12d3-a456-426614174001"
          },
          "name": {
            "type": "string",
            "example": "Margherita Pizza"
          },
          "price": {
            "type": "number",
            "example": 12.99
          },
          "currency": {
            "type": "string",
            "example": "EUR"
          },
          "quantity": {
            "type": "number",
            "example": 2
          }
        },
        "required": [
          "menuItemId",
          "restaurantId",
          "name",
          "price",
          "currency",
          "quantity"
        ]
      },
      "CartResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "123e4567-e89b-12d3-a456-426614174000"
          },
          "customerId": {
            "type": "string",
            "example": "123e4567-e89b-12d3-a456-426614174001"
          },
          "restaurantId": {
            "type": "object",
            "example": "123e4567-e89b-12d3-a456-426614174002"
          },
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CartItemDto"
            }
          },
          "totalAmount": {
            "type": "number",
            "example": 25.98
          },
          "currency": {
            "type": "string",
            "example": "EUR"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string",
            "example": "2024-01-01T00:00:00.000Z"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string",
            "example": "2024-01-01T00:00:00.000Z"
          }
        },
        "required": [
          "id",
          "customerId",
          "items",
          "totalAmount",
          "currency",
          "createdAt",
          "updatedAt"
        ]
      },
      "AddItemToCartDto": {
        "type": "object",
        "properties": {
          "menuItemId": {
            "type": "string",
            "format": "uuid"
          },
          "restaurantId": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string",
            "minLength": 1
          },
          "price": {
            "type": "number",
            "minimum": 0,
            "exclusiveMinimum": true
          },
          "currency": {
            "type": "string",
            "default": "EUR"
          },
          "quantity": {
            "type": "integer",
            "minimum": 0,
            "exclusiveMinimum": true,
            "default": 1
          }
        },
        "required": [
          "menuItemId",
          "restaurantId",
          "name",
          "price"
        ]
      },
      "UpdateItemQuantityDto": {
        "type": "object",
        "properties": {
          "menuItemId": {
            "type": "string",
            "format": "uuid"
          },
          "quantity": {
            "type": "integer",
            "minimum": 0,
            "exclusiveMinimum": false
          }
        },
        "required": [
          "menuItemId",
          "quantity"
        ]
      },
      "RemoveItemFromCartDto": {
        "type": "object",
        "properties": {
          "menuItemId": {
            "type": "string",
            "format": "uuid"
          }
        },
        "required": [
          "menuItemId"
        ]
      },
      "CheckoutCartDto": {
        "type": "object",
        "properties": {
          "deliveryAddress": {
            "type": "object",
            "properties": {
              "street": {
                "type": "string",
                "minLength": 1
              },
              "city": {
                "type": "string",
                "minLength": 1
              },
              "postalCode": {
                "type": "string",
                "minLength": 1
              },
              "country": {
                "type": "string",
                "minLength": 1
              },
              "latitude": {
                "type": "number"
              },
              "longitude": {
                "type": "number"
              },
              "instructions": {
                "type": "string"
              }
            },
            "required": [
              "street",
              "city",
              "postalCode",
              "country"
            ]
          },
          "deliveryFee": {
            "type": "number",
            "minimum": 0,
            "exclusiveMinimum": false,
            "default": 0
          }
        },
        "required": [
          "deliveryAddress"
        ]
      },
...
    }
  }
}